Bug+GTree
- Fix Some Bugs
- Add Growing Tree Algorithm
