--- README.md
@@ -1,5 +1,0 @@
-# MazeProject
-# A game That use Algorithms to generate Dynamic Mazes
-# Used Generation Algorithm : SideWinder,Prim And Growing Binary Tree
-# Solving Algorithm : Backtracking
-# With C++ And Qt.
