+++ include/SideWinder.h
@@ -1,0 +1,18 @@
+#ifndef SIDEWINDER_H
+#define SIDEWINDER_H
+
+#include "Maze.h"
+
+
+class SideWinder : public Maze
+{
+    public:
+        SideWinder(int,int);
+        ~SideWinder();
+
+    protected:
+
+    private:
+};
+
+#endif // SIDEWINDER_H
