+++ include/growing_tree.h
@@ -1,0 +1,21 @@
+#ifndef GROWING_TREE_H
+#define GROWING_TREE_H
+
+#include "Maze.h"
+#include <bits/stdc++.h>
+
+using namespace std;
+
+class growing_tree : public Maze
+{
+    public:
+        growing_tree(int,int);
+        direction int_to_Dir(int);
+        ~growing_tree();
+
+    protected:
+
+    private:
+};
+
+#endif // GROWING_TREE_H
